Json画面のSeenOnをアイコンで表示に
